+++ app/src/stores/errorStore/error-store.js
@@ -1,0 +1,32 @@
+import { observable, action, makeObservable } from 'mobx';
+
+class ErrorStore {
+  authErrors = observable.map();
+
+  setError = (code, error) => {
+    if (code >= 100 && code < 199) {
+      this.authErrors.set(code, error);
+    }
+  };
+
+  resetAll = () => {
+    this.authErrors.clear();
+  };
+
+  reset = (codes) => {
+    codes.forEach((code) => this.authErrors.delete(code));
+  };
+
+  constructor() {
+    makeObservable(this, {
+      authErrors: observable,
+      setError: action,
+      resetAll: action,
+      reset: action,
+    });
+  }
+}
+
+export const errorStore = new ErrorStore();
+
+export { ErrorStore };
