+++ app/src/stores/index.ts
@@ -1,9 +1,10 @@
-import dataStore from './dataStore';
-import nodeStore from './nodeStore';
-import modalStore from './modalStore';
-import defStore from './defStore';
-import errorStore from './errorStore';
+import { dataStore } from './dataStore';
+import { nodeStore } from './nodeStore';
+import { modalStore } from './modalStore';
+import { defStore } from './defStore';
+import { errorStore } from './errorStore';
 
+// TODO: Add StoreTypes
 const stores = {
   dataStore,
   nodeStore,
